DEMRUM-5474: Move CollectingSpanExporter to Testing Support folder
