+++ SplunkNavigation/Tests/SplunkNavigationTests/Testing Support/CollectingSpanExporter.swift
@@ -1,0 +1,43 @@
+//
+/*
+Copyright 2026 Splunk Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import OpenTelemetryApi
+import OpenTelemetrySdk
+
+final class CollectingSpanExporter: SpanExporter {
+    private let lock = NSLock()
+    private var collected: [SpanData] = []
+
+    var spans: [SpanData] {
+        lock.withLock { collected }
+    }
+
+    func reset() {
+        lock.withLock { collected.removeAll() }
+    }
+
+    func export(spans: [SpanData], explicitTimeout _: TimeInterval?) -> SpanExporterResultCode {
+        lock.withLock { collected.append(contentsOf: spans) }
+        return .success
+    }
+
+    func flush(explicitTimeout _: TimeInterval?) -> SpanExporterResultCode {
+        .success
+    }
+
+    func shutdown(explicitTimeout _: TimeInterval?) {}
+}
